tasks

+++ main.h
@@ -7,7 +7,7 @@
 #include <limits.h>
 #include <stdlib.h>
 int _printf(const char *foramt, ...);
-void _print_number(int);
+void _print_number(long int);
 int wrchar(int);
 int lengnum(int);
 int len_bin(unsigned int);
